fix(ui): prevent PresetBrowserView crash on editor close
Fixed use-after-free crash when closing plugin window after opening
preset browser. Root cause: destructor called unregisterTextEditListener()
on child view already destroyed by VSTGUI's container cleanup.

VSTGUI's CViewContainer destroys child views via beforeDelete()->removeAll()
BEFORE the parent destructor runs. Solution: removed the listener
unregistration call since child views are already gone.

Also includes:
- ASan build fixes for MSVC (/RTC1, /ZI conflicts)
- VST-GUIDE.md section on CViewContainer child destruction order
- intptr_t fix for C4312 warning in editor_lifecycle_test.cpp

+++ plugins/iterum/tests/unit/vst/editor_lifecycle_test.cpp
@@ -34,6 +34,8 @@
 
 #include <catch2/catch_test_macros.hpp>
 
+#include <cstdint>
+
 // ==============================================================================
 // PATTERN TEST: Direct Pointer vs Indirect Pointer
 // ==============================================================================
@@ -44,7 +46,7 @@
 // Simulates a simplified VST3Editor for testing
 struct MockEditor {
     bool isValid = true;
-    int frameId = 0;  // Simulates getFrame() returning different objects
+    intptr_t frameId = 0;  // Simulates getFrame() returning different objects
 
     // Simulate getFrame() - returns nullptr if editor was destroyed
     void* getFrame() { return isValid ? reinterpret_cast<void*>(frameId) : nullptr; }
